HP-2008-01-16-1: Bugfix:    Fixed broken build on irix65 after checkin HP-2008-01-16-0                  Changed:   sge_ijs_comm.c, sge_shepherd_ijs.c

+++ source/daemons/shepherd/sge_shepherd_ijs.c
@@ -151,7 +151,7 @@ static int messages_in_send_queue(COMMUNICATION_HANDLE *handle)
       }
       cl_raw_list_unlock(handle->connection_list);
    } 
-   return elems;
+   return (int)elems;
 }
 
 /****** append_to_buf() ******************************************************
